NIH_SimAmp_Quan/cGAN/requirements.txt
torch>=0.4.0
torchvision
torchaudio
matplotlib
numpy
scipy
pillow
urllib3
scikit-image
librosa
tqdm
speechbrain